#270: Move files to create package
