+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/application/webservice/schema/create/BlueCardApplication.kt
@@ -1,4 +1,4 @@
-package app.ehrenamtskarte.backend.application.webservice.schema
+package app.ehrenamtskarte.backend.application.webservice.schema.create
 
 data class BlueEakCardApplication(
     val personalData: PersonalData,
